Moved StatisticsDatabase class in a different file + /tmp/statistics path by default.
